+++ src/Settings/SettingsManager.cpp
@@ -30,6 +30,9 @@
 QVariantMap *SettingsManager::cur = nullptr;
 QVariantMap *SettingsManager::def = nullptr;
 QMap<QString, QString> *SettingsManager::settingPath = nullptr;
+QMap<QString, QString> *SettingsManager::settingTrPath = nullptr;
+QMap<QString, QString> *SettingsManager::pathSetting = nullptr;
+QMap<QString, QWidget *> *SettingsManager::settingWidget = nullptr;
 
 const static QStringList configFileLocations = {
 #ifdef PORTABLE_VERSION
@@ -97,10 +100,19 @@ void SettingsManager::init()
         delete def;
     if (settingPath)
         delete settingPath;
+    if (settingTrPath)
+        delete settingTrPath;
+    if (pathSetting)
+        delete pathSetting;
+    if (settingWidget)
+        delete settingWidget;
 
     cur = new QVariantMap();
     def = new QVariantMap();
     settingPath = new QMap<QString, QString>();
+    settingTrPath = new QMap<QString, QString>();
+    pathSetting = new QMap<QString, QString>();
+    settingWidget = new QMap<QString, QWidget *>();
 
     generateDefaultSettings();
 
@@ -113,11 +125,21 @@ void SettingsManager::deinit()
 {
     saveSettings(Util::configFilePath(configFileLocations[0]));
 
-    delete cur;
-    delete def;
-    delete settingPath;
+    if (cur)
+        delete cur;
+    if (def)
+        delete def;
+    if (settingPath)
+        delete settingPath;
+    if (settingTrPath)
+        delete settingTrPath;
+    if (pathSetting)
+        delete pathSetting;
+    if (settingWidget)
+        delete settingWidget;
     cur = def = nullptr;
-    settingPath = nullptr;
+    settingPath = settingTrPath = pathSetting = nullptr;
+    settingWidget = nullptr;
 }
 
 void SettingsManager::generateDefaultSettings()
@@ -199,9 +221,11 @@ void SettingsManager::reset()
     *cur = *def;
 }
 
-void SettingsManager::setPath(const QString &key, const QString &path)
+void SettingsManager::setPath(const QString &key, const QString &path, const QString &trPath)
 {
     settingPath->insert(key, path);
+    settingTrPath->insert(key, trPath);
+    pathSetting->insert(path, key);
 }
 
 QString SettingsManager::getPathText(const QString &key, bool parent)
@@ -209,19 +233,44 @@ QString SettingsManager::getPathText(const QString &key, bool parent)
     if (!settingPath->contains(key))
     {
 #ifdef QT_DEBUG
-        qDebug() << "SettingsManager: Getting unknown key path: " << key;
+        qDebug() << "SettingsManager: Getting unknown path: " << key;
 #endif
-        LOG_WARN("Getting unknown key path: " << key);
+        LOG_WARN("Getting unknown path: " << INFO_OF(key));
         return "Unknown";
     }
-    auto nodes = settingPath->value(key).split('/');
-    for (int i = 0; i < nodes.count() - 1; ++i)
-        nodes[i] = QCoreApplication::translate("PreferencesWindow", nodes[i].toStdString().c_str());
+    auto path = settingPath->value(key);
+    auto trPath = QCoreApplication::translate("PreferencesWindow", "Preferences") + "/" + settingTrPath->value(key);
     if (parent)
-        nodes.pop_back();
-    else
-        nodes.back() = QCoreApplication::translate("SettingsInfo", nodes.back().toStdString().c_str());
-    return nodes.join("->");
+    {
+        path.chop(path.length() - path.lastIndexOf('/'));
+        trPath.chop(trPath.length() - trPath.lastIndexOf('/'));
+    }
+    return QString("<a href='#Preferences/%1'>%2</a>").arg(path).arg(trPath.replace('/', "->"));
+}
+
+QString SettingsManager::getKeyOfPath(const QString &path)
+{
+    if (!pathSetting->contains(path))
+    {
+        LOG_WARN("Getting unknown key of path: " << INFO_OF(path));
+        return QString();
+    }
+    return pathSetting->value(path);
+}
+
+void SettingsManager::setWidget(const QString &key, QWidget *widget)
+{
+    settingWidget->insert(key, widget);
+}
+
+QWidget *SettingsManager::getWidget(const QString &key)
+{
+    if (!settingWidget->contains(key))
+    {
+        LOG_WARN("Getting unknown widget: " << INFO_OF(key));
+        return nullptr;
+    }
+    return settingWidget->value(key);
 }
 
 QStringList SettingsManager::keyStartsWith(const QString &head)