+++ src/mainwindow.cpp
@@ -59,15 +59,26 @@ MainWindow::MainWindow(int index, AppWindow *window, QWidget *parent)
       autoSaveTimer(new QTimer(this))
 {
     LOG_INFO(INFO_OF(index));
+
     ui->setupUi(this);
-    setupCore();
-    setTestCases();
+
+    log = new MessageLogger(window->getPreferencesWindow(), this);
+    ui->messageLoggerLayout->addWidget(log);
+
+    formatter = new Extensions::ClangFormatter(SettingsHelper::getClangFormatPath(),
+                                               SettingsHelper::getClangFormatStyle(), log);
+
+    testcases = new Widgets::TestCases(log, this);
+    ui->testCasesLayout->addWidget(testcases);
+    connect(testcases, SIGNAL(checkerChanged()), this, SLOT(updateChecker()));
+    connect(testcases, SIGNAL(requestRun(int)), this, SLOT(runTestCase(int)));
+
     setEditor();
     connect(fileWatcher, SIGNAL(fileChanged(const QString &)), this, SLOT(onFileWatcherChanged(const QString &)));
     connect(
         autoSaveTimer, &QTimer::timeout, autoSaveTimer, [this] { saveFile(AutoSave, tr("Auto Save"), false); },
         Qt::DirectConnection);
-    applySettings("", true);
+    applySettings("");
     QTimer::singleShot(0, [this] { setLanguage(language); }); // See issue #187 for more information
 }
 
@@ -113,16 +124,6 @@ MainWindow::~MainWindow()
     delete log;
 }
 
-// ************************* RAII HELPER *****************************
-
-void MainWindow::setTestCases()
-{
-    testcases = new Widgets::TestCases(log, this);
-    ui->testCasesLayout->addWidget(testcases);
-    connect(testcases, SIGNAL(checkerChanged()), this, SLOT(updateChecker()));
-    connect(testcases, SIGNAL(requestRun(int)), this, SLOT(runTestCase(int)));
-}
-
 void MainWindow::setEditor()
 {
     editor = new QCodeEditor();
@@ -139,14 +140,6 @@ void MainWindow::setEditor()
     // a selection (and the cursor is at the begin of the selection)
     connect(editor, SIGNAL(cursorPositionChanged()), this, SLOT(updateCursorInfo()));
     connect(editor, SIGNAL(selectionChanged()), this, SLOT(updateCursorInfo()));
-}
-
-void MainWindow::setupCore()
-{
-    log = new MessageLogger();
-    log->setContainer(ui->compilerEdit);
-    formatter = new Extensions::ClangFormatter(SettingsHelper::getClangFormatPath(),
-                                               SettingsHelper::getClangFormatStyle(), log);
 }
 
 void MainWindow::compile()
@@ -176,6 +169,7 @@ void MainWindow::compile()
     connect(compiler, SIGNAL(compilationFinished(const QString &)), this, SLOT(onCompilationFinished(const QString &)));
     connect(compiler, SIGNAL(compilationErrorOccurred(const QString &)), this,
             SLOT(onCompilationErrorOccurred(const QString &)));
+    connect(compiler, SIGNAL(compilationFailed(const QString &)), this, SLOT(onCompilationFailed(const QString &)));
     connect(compiler, SIGNAL(compilationKilled()), this, SLOT(onCompilationKilled()));
     compiler->start(path, filePath, compileCommand(), language);
 }
@@ -291,8 +285,10 @@ void MainWindow::setCFToolUI()
     {
         submitToCodeforces->setEnabled(false);
         log->error(tr("CF Tool"),
-                   tr("You will not be able to submit code to Codeforces because CF Tool is not installed or is "
-                      "not on SYSTEM PATH. You can set it manually in settings."));
+                   tr("You need to install CF Tool to submit your code to Codeforces. If already installed, you can "
+                      "add it in the PATH environment variable or check your settings at %1.")
+                       .arg(SettingsHelper::pathOfCFPath()),
+                   false);
     }
 }
 
@@ -537,7 +533,8 @@ void MainWindow::applyCompanion(const Extensions::CompanionData &data)
             {
                 log->error("Companion",
                            tr("Unknown attribute: [%1]. Please check the head comments setting at %2.")
-                               .arg(path.join('.').arg(SettingsHelper::pathOfCompetitiveCompanionHeadComments())));
+                               .arg(path.join('.').arg(SettingsHelper::pathOfCompetitiveCompanionHeadComments())),
+                           false);
             }
             else if (value.isObject() || value.isArray())
             {
@@ -585,9 +582,9 @@ void MainWindow::applyCompanion(const Extensions::CompanionData &data)
         customTimeLimit = data.timeLimit;
 }
 
-void MainWindow::applySettings(const QString &pagePath, bool shouldPerformDigonistic)
+void MainWindow::applySettings(const QString &pagePath)
 {
-    LOG_INFO(INFO_OF(pagePath) << BOOL_INFO_OF(shouldPerformDigonistic));
+    LOG_INFO(INFO_OF(pagePath));
 
     auto pageChanged = [pagePath](const QString &page) { return pagePath.isEmpty() || pagePath == page; };
 
@@ -643,11 +640,6 @@ void MainWindow::applySettings(const QString &pagePath, bool shouldPerformDigoni
         setLanguage(SettingsHelper::getDefaultLanguage());
     }
 
-    if (shouldPerformDigonistic && (pageChanged(QStringLiteral("Language/%1/%1 Commands").arg(language))))
-    {
-        performCompileAndRunDiagonistics();
-    }
-
     if (pageChanged("Appearance/General"))
     {
         testcases->updateHeights();
@@ -665,7 +657,7 @@ void MainWindow::applySettings(const QString &pagePath, bool shouldPerformDigoni
 
     if (pageChanged("Appearance/Font"))
     {
-        ui->compilerEdit->setFont(SettingsHelper::getMessageLoggerFont());
+        log->setFont(SettingsHelper::getMessageLoggerFont());
         testcases->setTestCaseEditFont(SettingsHelper::getTestCasesFont());
     }
 
@@ -773,7 +765,6 @@ void MainWindow::setLanguage(const QString &lang)
     Util::applySettingsToEditor(editor, language);
     customCompileCommand.clear();
     ui->changeLanguageButton->setText(language);
-    performCompileAndRunDiagonistics();
     isLanguageSet = true;
     emit editorLanguageChanged(this);
 }
@@ -926,7 +917,8 @@ void MainWindow::loadFile(const QString &loadPath)
                       "open file length limit at %3.")
                        .arg(path)
                        .arg(SettingsHelper::getOpenFileLengthLimit())
-                       .arg(SettingsHelper::pathOfOpenFileLengthLimit()));
+                       .arg(SettingsHelper::pathOfOpenFileLengthLimit()),
+                   false);
         setText("");
         setFilePath("");
         return;
@@ -1320,27 +1312,6 @@ QSplitter *MainWindow::getRightSplitter()
     return ui->rightSplitter;
 }
 
-void MainWindow::performCompileAndRunDiagonistics()
-{
-    bool compilerResult = true;
-    bool runResult = true;
-
-    if (language == "C++" || language == "Java")
-        compilerResult =
-            Core::Compiler::check(SettingsManager::get(QString("%1/Compile Command").arg(language)).toString());
-
-    if (language == "Java" || language == "Python")
-        runResult = Core::Compiler::check(SettingsManager::get(QString("%1/Run Command").arg(language)).toString());
-
-    if (!compilerResult)
-        log->error(tr("Compiler"),
-                   tr("The compile command for %1 is invalid. Is the compiler in the system PATH?").arg(language));
-
-    if (!runResult)
-        log->error(tr("Runner"),
-                   tr("The run command for %1 is invalid. Is the runner in the system Path?").arg(language));
-}
-
 QString MainWindow::compileCommand() const
 {
     if (customCompileCommand.isEmpty())
@@ -1414,9 +1385,15 @@ void MainWindow::onCompilationErrorOccurred(const QString &error)
             log->warn(
                 tr("Compile Errors"),
                 tr("Have you set a proper name for the main class in your solution? If not, you can set it at %1.")
-                    .arg(SettingsHelper::pathOfJavaClassName()));
+                    .arg(SettingsHelper::pathOfJavaClassName()),
+                false);
         }
     }
+}
+
+void MainWindow::onCompilationFailed(const QString &reason)
+{
+    log->error(tr("Compiler"), tr("Failed to start compilation: %1").arg(reason), false);
 }
 
 void MainWindow::onCompilationKilled()
@@ -1474,7 +1451,7 @@ void MainWindow::onRunFinished(int index, const QString &out, const QString &err
 
 void MainWindow::onFailedToStartRun(int index, const QString &error)
 {
-    log->error(getRunnerHead(index), error);
+    log->error(getRunnerHead(index), error, false);
 }
 
 void MainWindow::onRunOutputLimitExceeded(int index, const QString &type)
@@ -1486,7 +1463,8 @@ void MainWindow::onRunOutputLimitExceeded(int index, const QString &type)
             .arg(type)
             .arg(index + 1)
             .arg(SettingsHelper::getOutputLengthLimit())
-            .arg(SettingsHelper::pathOfOutputLengthLimit()));
+            .arg(SettingsHelper::pathOfOutputLengthLimit()),
+        false);
 }
 
 void MainWindow::onRunKilled(int index)